+++ mem_dbg/tests/test_atomics.rs
@@ -7,6 +7,7 @@ use std::sync::atomic::*;
 #[test]
 fn test_atomic_bool_in_struct() {
     #[derive(MemSize)]
+    #[move_type]
     struct Test {
         atomic: AtomicBool,
     }
@@ -22,6 +23,7 @@ fn test_atomic_bool_in_struct() {
 #[test]
 fn test_atomic_integers_in_struct() {
     #[derive(MemSize)]
+    #[move_type]
     struct Test {
         atomic_i8: AtomicI8,
         atomic_i16: AtomicI16,
@@ -57,6 +59,7 @@ fn test_atomic_integers_in_struct() {
 #[test]
 fn test_atomic_with_memdbg() {
     #[derive(MemSize, MemDbg)]
+    #[move_type]
     struct Test {
         atomic_bool: AtomicBool,
         atomic_u32: AtomicU32,