+++ mem_dbg/tests/test_unsafe_cell.rs
@@ -8,6 +8,7 @@ use std::cell::Cell;
 #[test]
 fn test_unsafe_cell_in_struct() {
     #[derive(MemSize, MemDbg)]
+    #[move_type]
     struct Test {
         field: UnsafeCell<i32>,
     }
@@ -25,6 +26,7 @@ fn test_unsafe_cell_in_struct() {
 #[test]
 fn test_cell_in_struct() {
     #[derive(MemSize, MemDbg)]
+    #[move_type]
     struct Test {
         field: Cell<i32>,
     }